+++ app/controllers/quizzes_controller.rb
@@ -2,13 +2,14 @@
 
 class QuizzesController < ApplicationController
   before_action :authenticate_user_using_x_auth_token, except: %i[check_slug]
-  before_action :load_quiz, only: %i[show update destroy]
+  before_action :load_quiz, except: %i[index create check_slug]
   def index
     @quizzes = @current_user.quizzes.order("created_at DESC").as_json(only: %i[id title ])
   end
 
   def create
     quiz = @current_user.quizzes.new(quiz_params)
+    authorize quiz
     if quiz.save
       render status: :ok, json: { notice: t("successfully_created") }
     else
@@ -18,10 +19,12 @@ class QuizzesController < ApplicationController
   end
 
   def show
+    authorize @quiz
     @questions = @quiz.questions.includes(:options)
   end
 
   def update
+    authorize @quiz
     if !quiz_params[:setslug]
       if @quiz.update(title: quiz_params[:title])
         render status: :ok, json: { notice: t("successfully_updated", entity: "Quiz") }
@@ -40,11 +43,16 @@ class QuizzesController < ApplicationController
   end
 
   def destroy
+    authorize @quiz
     if @quiz.destroy
       render status: :ok, json: { notice: t("deleted_successfully") }
     else
       render status: :unprocessable_entity, json: { error: @quiz.errors.full_messages.to_sentence }
     end
+  end
+
+  def retrieve_title
+    authorize @quiz
   end
 
   def check_slug
@@ -60,7 +68,7 @@ class QuizzesController < ApplicationController
     end
 
     def load_quiz
-      @quiz = @current_user.quizzes.find_by(id: params[:id])
+      @quiz = Quiz.find_by(id: params[:id])
       unless @quiz
         render status: :not_found, json: { error: t("not_found", entity: "Quiz") }
       end
